more cleanup

+++ cocos/3d/CCMesh.h
@@ -52,9 +52,9 @@ public:
     /**create mesh from positions, normals, and so on, sigle SubMesh*/
     static Mesh* create(const std::vector<float>& positions, const std::vector<float>& normals, const std::vector<float>& texs, const IndexArray& indices);
     /**create mesh with vertex attributes*/
-    CC_DEPRECATED(v3) static Mesh* create(const std::vector<float>& vertices, int perVertexSizeInFloat, const IndexArray& indices, int numIndex, const std::vector<MeshVertexAttrib>& attribs, int attribCount){ return create(vertices, perVertexSizeInFloat, indices, attribs); }
+    CC_DEPRECATED(v3) static Mesh* create(const std::vector<float>& vertices, ssize_t perVertexSizeInFloat, const IndexArray& indices, int numIndex, const std::vector<MeshVertexAttrib>& attribs, int attribCount){ return create(vertices, perVertexSizeInFloat, indices, attribs); }
     
-    static Mesh* create(const std::vector<float>& vertices, int perVertexSizeInFloat, const IndexArray& indices, const std::vector<MeshVertexAttrib>& attribs);
+    static Mesh* create(const std::vector<float>& vertices, ssize_t perVertexSizeInFloat, const IndexArray& indices, const std::vector<MeshVertexAttrib>& attribs);
     
     /** create mesh */
     static Mesh* create(const std::string& name, MeshIndexData* indexData, MeshSkin* skin = nullptr);
@@ -68,7 +68,7 @@ public:
     /**get MeshVertexAttribute by index*/
     const MeshVertexAttrib& getMeshVertexAttribute(int idx);
     /**get per vertex size in bytes*/
-    int getVertexSizeInBytes() const;
+    ssize_t getVertexSizeInBytes() const;
 
     /**texture getter and setter*/
     void setTexture(const std::string& texPath);
